requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
beautifulsoup4>=4.12,<5.0
python-docx>=1.1,<2.0
requests>=2.31,<3.0
